+++ lib/ruby-handlebars/helpers/helper_missing_helper.rb
@@ -1,0 +1,14 @@
+module Handlebars
+  class UnknownHelper < StandardError
+  end
+
+  module Helpers
+    class HelperMissingHelper
+      def self.register(hbs)
+        hbs.register_helper('helperMissing') do |context, name|
+          raise(::Handlebars::UnknownHelper, "Helper \"#{name}\" does not exist" )
+        end
+      end
+    end
+  end
+end